# olympapi-validation-suite/docs/COLLECTIONS/OlympAPI Test Server- swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OlympAPI Test Server",
        "version": "1.0.0",
        "description": "A dedicated test server covering all OlympAPI features: HTTP methods, auth, body types, chaining, scripts, SSL, and proxy."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "HTTP"
        },
        {
            "url": "https://localhost:3443",
            "description": "HTTPS (self-signed)"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "Use \"test-token-123\" for /auth/bearer, or the token from POST /auth/login for /auth/me and /users/:id"
            },
            "BasicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "Use admin:password123"
            },
            "ApiKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "Use \"api-key-xyz\""
            },
            "ApiKeyQuery": {
                "type": "apiKey",
                "in": "query",
                "name": "api_key",
                "description": "Use \"api-key-xyz\""
            }
        }
    },
...
    "paths": {
        "/auth/bearer": {
            "get": {
                "summary": "Protected by Bearer token",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authenticated successfully"
                    },
                    "401": {
                        "description": "Invalid or missing token"
                    }
                }
            }
        },
        "/auth/basic": {
            "get": {
                "summary": "Protected by HTTP Basic auth (admin:password123)",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authenticated successfully"
                    },
                    "401": {
                        "description": "Invalid or missing credentials"
                    }
                }
            }
        },
        "/auth/apikey": {
            "get": {
                "summary": "Protected by API key (header or query param)",
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-API-Key",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "api_key",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authenticated successfully"
                    },
                    "401": {
                        "description": "Missing or invalid API key"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login with username and password",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "demo"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "demo123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful, returns token"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "summary": "Get current user profile (requires login token)",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile"
                    },
                    "401": {
                        "description": "Invalid or missing token"
                    }
                }
            }
        },
        "/body/json": {
            "post": {
                "summary": "Parse and echo a JSON body",
                "tags": [
                    "Body"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Echoed JSON body info"
                    },
                    "400": {
                        "description": "Body is not valid JSON"
                    }
                }
            }
        },
        "/body/form": {
            "post": {
                "summary": "Parse and echo a URL-encoded form body",
                "tags": [
                    "Body"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Echoed form fields"
                    }
                }
            }
        },
        "/body/validate": {
            "post": {
                "summary": "Validate required fields name and email",
                "tags": [
                    "Body"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation passed"
                    },
                    "422": {
                        "description": "Missing required fields"
                    }
                }
            }
        },
        "/body/multipart": {
            "post": {
                "summary": "Accept multipart/form-data (text fields + file uploads)",
                "tags": [
                    "Body"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Echoed fields and file metadata"
                    }
                }
            }
        },
        "/body/binary": {
            "post": {
                "summary": "Accept a raw binary body (any content type, max 10 MB)",
                "tags": [
                    "Body"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns size and content type"
                    }
                }
            }
        },
        "/graphql": {
            "post": {
                "summary": "Mock GraphQL endpoint — echoes query and variables",
                "tags": [
                    "Body"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string"
                                    },
                                    "variables": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "GraphQL echo response"
                    }
                }
            }
        },
...
    }
}